+++ nodes/Fireflies/resources/aiApp.properties.ts
