+++ nodes/Fireflies/resources/audio.fields.ts
