+++ nodes/Fireflies/resources/audio.properties.ts
