+++ nodes/Fireflies/resources/transcript.fields.ts
